src/tools/rotate/pyproject.toml
[tool.poetry]
name = "picturas-rotate-tool-ms"
version = "0.1.0"
description = "PictuRAS microservice (tool) that rotates images"
authors = ["Francisco Lameirão <kikofallout@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.13"
pika = "^1.3.2"
pydantic = "^2.10.3"
pillow = "^11.0.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
